Refactor colour node input
